fix: Add validation to ensure only one "id" parameter is allowed for user deletion

+++ src/modules/sync-server/entities/user/user.controller.ts
@@ -60,6 +60,11 @@ export class UserController {
   @Delete()
   private async deleteUser(req: Request): Promise<UserWithoutPasswordDTO> {
     const { id } = req.params
+
+    if (Array.isArray(id)) {
+      throw new BadRequestException('Only one "id" parameter is allowed')
+    }
+
     const deletedUser = await this.userService.deleteUser(id)
     return UserMapper.fromEntityToDtoWithoutPassword(deletedUser)
   }
